appPyBills/requirements.txt
Flask==0.10.1
PyMySQL==0.6.7
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-WTF==0.12
Flask-Migrate==1.6.0
flask-bcrypt
python-slugify==1.1.4
Flask-Uploads==0.2.0